Added Guards

+++ src/utils/user.ts
@@ -3,7 +3,8 @@ import { setAccessToken } from "../store/actions/auth.actions";
 import { setSnackbarEvent } from "../store/actions/user.actions";
 
 export const logOutUser = () => {
-    localStorage.clear();
+    if (!localStorage.getItem("user")) return; 
     store.dispatch(setSnackbarEvent({ content: "Logged Out", variant: "info" }))
     store.dispatch(setAccessToken({ accessToken: '' }));
+    localStorage.clear();
 };
